Refactor statusbar clock text size
